moved NetworkState to a separate module
