+++ chain/network/src/peer_manager/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod connection;
+pub(crate) mod network_state;
 pub(crate) mod peer_manager_actor;
 pub(crate) mod peer_store;
