+++ chain/network/src/peer_manager/peer_manager_actor.rs
@@ -1,9 +1,7 @@
-use crate::accounts_data;
-use crate::concurrency::demux;
 use crate::config;
 use crate::network_protocol::{AccountData, SyncAccountsData};
-use crate::peer::peer_actor::{PeerActor, StreamConfig};
 use crate::peer_manager::connection;
+use crate::peer_manager::network_state::NetworkState;
 use crate::peer_manager::peer_store::PeerStore;
 use crate::private_actix::{
     PeerRequestResult, PeersRequest, RegisterPeer, RegisterPeerResponse, StopMsg, Unregister,
@@ -16,7 +14,7 @@ use crate::routing::routing_table_view::RoutingTableView;
 use crate::stats::metrics;
 use crate::store;
 use crate::types::{
-    ChainInfo, ConnectedPeerInfo, FullPeerInfo, GetNetworkInfo, NetworkClientMessages, NetworkInfo,
+    ConnectedPeerInfo, FullPeerInfo, GetNetworkInfo, NetworkClientMessages, NetworkInfo,
     NetworkRequests, NetworkResponses, PeerManagerMessageRequest, PeerManagerMessageResponse,
     PeerMessage, RoutingTableUpdate, SetChainInfo,
 };
@@ -26,17 +24,15 @@ use actix::{
 };
 use anyhow::bail;
 use anyhow::Context as _;
-use arc_swap::ArcSwap;
 use near_network_primitives::time;
 use near_network_primitives::types::{
     AccountOrPeerIdOrHash, Ban, Edge, KnownPeerStatus, KnownProducer, NetworkViewClientMessages,
-    NetworkViewClientResponses, OutboundTcpConnect, PeerIdOrHash, PeerInfo, PeerType, Ping, Pong,
-    RawRoutedMessage, ReasonForBan, RoutedMessageBody, RoutedMessageV2, StateResponseInfo,
+    NetworkViewClientResponses, OutboundTcpConnect, PeerInfo, PeerType, Ping, Pong,
+    RawRoutedMessage, ReasonForBan, RoutedMessageBody, StateResponseInfo,
 };
 use near_network_primitives::types::{EdgeState, PartialEdgeInfo};
 use near_performance_metrics_macros::perf;
 use near_primitives::block::GenesisId;
-use near_primitives::hash::CryptoHash;
 use near_primitives::network::{AnnounceAccount, PeerId};
 use near_primitives::types::{AccountId, EpochId};
 use near_rate_limiter::{
@@ -48,13 +44,11 @@ use rand::thread_rng;
 use std::cmp::{max, min};
 use std::collections::{HashMap, HashSet};
 use std::net::SocketAddr;
-use std::sync::atomic::{AtomicUsize, Ordering};
+use std::sync::atomic::Ordering;
 use std::sync::Arc;
-use tokio::net::{TcpListener, TcpStream};
+use tokio::net::TcpListener;
 use tracing::{debug, error, info, trace, warn, Instrument};
 
-/// How often to request peers from active peers.
-const REQUEST_PEERS_INTERVAL: time::Duration = time::Duration::milliseconds(60_000);
 /// How much time to wait (in milliseconds) after we send update nonce request before disconnecting.
 /// This number should be large to handle pair of nodes with high latency.
 const WAIT_ON_TRY_UPDATE_NONCE: time::Duration = time::Duration::milliseconds(6_000);
@@ -68,8 +62,6 @@ const EXPONENTIAL_BACKOFF_RATIO: f64 = 1.1;
 const MONITOR_PEERS_MAX_DURATION: time::Duration = time::Duration::milliseconds(60_000);
 /// The initial waiting time between consecutive attempts to establish connection
 const MONITOR_PEERS_INITIAL_DURATION: time::Duration = time::Duration::milliseconds(10);
-/// Limit number of pending Peer actors to avoid OOM.
-pub(crate) const LIMIT_PENDING_PEERS: usize = 60;
 /// How ofter should we broadcast edges.
 const BROADCAST_VALIDATED_EDGES_INTERVAL: time::Duration = time::Duration::milliseconds(50);
 /// Maximum amount of time spend processing edges.
@@ -89,7 +81,7 @@ const REPORT_BANDWIDTH_THRESHOLD_COUNT: usize = 10_000;
 /// How long a peer has to be unreachable, until we prune it from the in-memory graph.
 const PRUNE_UNREACHABLE_PEERS_AFTER: time::Duration = time::Duration::hours(1);
 
-// Remove the edges that were created more that this duration ago.
+/// Remove the edges that were created more that this duration ago.
 const PRUNE_EDGES_AFTER: time::Duration = time::Duration::minutes(30);
 // Don't accept nonces (edges) that are more than this delta from current time.
 // This value should be smaller than PRUNE_EDGES_AFTER (otherwise, the might accept the edge and garbage collect it seconds later).
@@ -99,11 +91,11 @@ const EDGE_NONCE_MAX_TIME_DELTA: time::Duration = time::Duration::minutes(20);
 /// We send these messages multiple times to reduce the chance that they are lost
 const IMPORTANT_MESSAGE_RESENT_COUNT: usize = 3;
 
-// If a peer is more than these blocks behind (comparing to our current head) - don't route any messages through it.
-// We are updating the list of unreliable peers every MONITOR_PEER_MAX_DURATION (60 seconds) - so the current
-// horizon value is roughly matching this threshold (if the node is 60 blocks behind, it will take it a while to recover).
-// If we set this horizon too low (for example 2 blocks) - we're risking excluding a lot of peers in case of a short
-// network issue.
+/// If a peer is more than these blocks behind (comparing to our current head) - don't route any messages through it.
+/// We are updating the list of unreliable peers every MONITOR_PEER_MAX_DURATION (60 seconds) - so the current
+/// horizon value is roughly matching this threshold (if the node is 60 blocks behind, it will take it a while to recover).
+/// If we set this horizon too low (for example 2 blocks) - we're risking excluding a lot of peers in case of a short
+/// network issue.
 const UNRELIABLE_PEER_HORIZON: u64 = 60;
 
 #[derive(Clone, PartialEq, Eq)]
@@ -125,220 +117,6 @@ impl TryFrom<&PeerInfo> for WhitelistNode {
             },
             account_id: pi.account_id.clone(),
         })
-    }
-}
-
-pub(crate) struct NetworkState {
-    /// PeerManager config.
-    pub config: Arc<config::VerifiedConfig>,
-    /// GenesisId of the chain.
-    pub genesis_id: GenesisId,
-    pub send_accounts_data_rl: demux::RateLimit,
-    /// Address of the client actor.
-    pub client_addr: Recipient<NetworkClientMessages>,
-    /// Address of the view client actor.
-    pub view_client_addr: Recipient<NetworkViewClientMessages>,
-    /// Address of the peer manager actor.
-    pub peer_manager_addr: Recipient<PeerToManagerMsg>,
-
-    /// Network-related info about the chain.
-    pub chain_info: ArcSwap<ChainInfo>,
-    /// AccountsData for TIER1 accounts.
-    pub accounts_data: Arc<accounts_data::Cache>,
-    /// Connected peers (inbound and outbound) with their full peer information.
-    pub tier2: connection::Pool,
-    /// Semaphore limiting inflight inbound handshakes.
-    pub inbound_handshake_permits: Arc<tokio::sync::Semaphore>,
-
-    /// View of the Routing table. It keeps:
-    /// - routing information - how to route messages
-    /// - edges adjacent to my_peer_id
-    /// - account id
-    /// Full routing table (that currently includes information about all edges in the graph) is now inside Routing Table.
-    pub routing_table_view: RoutingTableView,
-
-    /// Shared counter across all PeerActors, which counts number of `RoutedMessageBody::ForwardTx`
-    /// messages sincce last block.
-    pub txns_since_last_block: AtomicUsize,
-}
-
-impl NetworkState {
-    pub fn new(
-        config: Arc<config::VerifiedConfig>,
-        genesis_id: GenesisId,
-        client_addr: Recipient<NetworkClientMessages>,
-        view_client_addr: Recipient<NetworkViewClientMessages>,
-        peer_manager_addr: Recipient<PeerToManagerMsg>,
-        routing_table_view: RoutingTableView,
-        send_accounts_data_rl: demux::RateLimit,
-    ) -> Self {
-        Self {
-            genesis_id,
-            client_addr,
-            view_client_addr,
-            peer_manager_addr,
-            chain_info: Default::default(),
-            tier2: connection::Pool::new(config.node_id()),
-            accounts_data: Arc::new(accounts_data::Cache::new()),
-            routing_table_view,
-            send_accounts_data_rl,
-            inbound_handshake_permits: Arc::new(tokio::sync::Semaphore::new(LIMIT_PENDING_PEERS)),
-            config,
-            txns_since_last_block: AtomicUsize::new(0),
-        }
-    }
-
-    /// Query connected peers for more peers.
-    pub fn ask_for_more_peers(&self, clock: &time::Clock) {
-        let now = clock.now();
-        let msg = Arc::new(PeerMessage::PeersRequest);
-        for peer in self.tier2.load().ready.values() {
-            if now > peer.last_time_peer_requested.load() + REQUEST_PEERS_INTERVAL {
-                peer.send_message(msg.clone());
-            }
-        }
-    }
-
-    pub fn propose_edge(&self, peer1: &PeerId, with_nonce: Option<u64>) -> PartialEdgeInfo {
-        // When we create a new edge we increase the latest nonce by 2 in case we miss a removal
-        // proposal from our partner.
-        let nonce = with_nonce.unwrap_or_else(|| {
-            self.routing_table_view.get_local_edge(peer1).map_or(1, |edge| edge.next())
-        });
-        PartialEdgeInfo::new(&self.config.node_id(), peer1, nonce, &self.config.node_key)
-    }
-
-    /// Connects peer with given TcpStream.
-    /// It will fail (and log) if we have too many connections already,
-    /// or if the peer drops the connection in the meantime.
-    fn spawn_peer_actor(
-        self: &Arc<Self>,
-        clock: &time::Clock,
-        stream: TcpStream,
-        stream_cfg: StreamConfig,
-    ) {
-        if let Err(err) = PeerActor::spawn(clock.clone(), stream, stream_cfg, None, self.clone()) {
-            tracing::info!(target:"network", ?err, "PeerActor::spawn()");
-        };
-    }
-
-    pub async fn spawn_inbound(self: &Arc<Self>, clock: &time::Clock, stream: TcpStream) {
-        self.spawn_peer_actor(clock, stream, StreamConfig::Inbound);
-    }
-
-    pub async fn spawn_outbound(self: Arc<Self>, clock: time::Clock, peer_info: PeerInfo) {
-        let addr = match peer_info.addr {
-            Some(addr) => addr,
-            None => {
-                warn!(target: "network", ?peer_info, "Trying to connect to peer with no public address");
-                return;
-            }
-        };
-        // The `connect` may take several minutes. This happens when the
-        // `SYN` packet for establishing a TCP connection gets silently
-        // dropped, in which case the default TCP timeout is applied. That's
-        // too long for us, so we shorten it to one second.
-        //
-        // Why exactly a second? It was hard-coded in a library we used
-        // before, so we keep it to preserve behavior. Removing the timeout
-        // completely was observed to break stuff for real on the testnet.
-        let stream =
-            match tokio::time::timeout(std::time::Duration::from_secs(1), TcpStream::connect(addr))
-                .await
-            {
-                Ok(Ok(it)) => it,
-                Ok(Err(err)) => {
-                    info!(target: "network", ?addr, ?err, "Error connecting to");
-                    return;
-                }
-                Err(err) => {
-                    info!(target: "network", ?addr, ?err, "Error connecting to");
-                    return;
-                }
-            };
-        debug!(target: "network", ?peer_info, "Connecting");
-        self.spawn_peer_actor(&clock, stream, StreamConfig::Outbound { peer_id: peer_info.id });
-    }
-
-    // Determine if the given target is referring to us.
-    pub fn message_for_me(&self, target: &PeerIdOrHash) -> bool {
-        let my_peer_id = self.config.node_id();
-        match target {
-            PeerIdOrHash::PeerId(peer_id) => &my_peer_id == peer_id,
-            PeerIdOrHash::Hash(hash) => {
-                self.routing_table_view.compare_route_back(*hash, &my_peer_id)
-            }
-        }
-    }
-
-    pub fn send_ping(&self, clock: &time::Clock, nonce: u64, target: PeerId) {
-        let body = RoutedMessageBody::Ping(Ping { nonce, source: self.config.node_id() });
-        let msg = RawRoutedMessage { target: AccountOrPeerIdOrHash::PeerId(target), body };
-        self.send_message_to_peer(clock, msg);
-    }
-
-    pub fn send_pong(&self, clock: &time::Clock, nonce: u64, target: CryptoHash) {
-        let body = RoutedMessageBody::Pong(Pong { nonce, source: self.config.node_id() });
-        let msg = RawRoutedMessage { target: AccountOrPeerIdOrHash::Hash(target), body };
-        self.send_message_to_peer(clock, msg);
-    }
-
-    /// Route message to target peer.
-    /// Return whether the message is sent or not.
-    pub fn send_message_to_peer(&self, clock: &time::Clock, msg: RawRoutedMessage) -> bool {
-        self.send_signed_message_to_peer(
-            clock,
-            Box::new(msg.sign(
-                self.config.node_id(),
-                &self.config.node_key,
-                self.config.routed_message_ttl,
-                Some(clock.now_utc()),
-            )),
-        )
-    }
-
-    /// Route signed message to target peer.
-    /// Return whether the message is sent or not.
-    pub fn send_signed_message_to_peer(
-        &self,
-        clock: &time::Clock,
-        msg: Box<RoutedMessageV2>,
-    ) -> bool {
-        let my_peer_id = self.config.node_id();
-
-        // Check if the message is for myself and don't try to send it in that case.
-        if let PeerIdOrHash::PeerId(target) = &msg.msg.target {
-            if target == &my_peer_id {
-                debug!(target: "network", account_id = ?self.config.validator.as_ref().map(|v|v.account_id()), ?my_peer_id, ?msg, "Drop signed message to myself");
-                metrics::CONNECTED_TO_MYSELF.inc();
-                return false;
-            }
-        }
-
-        match self.routing_table_view.find_route(&clock, &msg.target) {
-            Ok(peer_id) => {
-                // Remember if we expect a response for this message.
-                if msg.msg.author == my_peer_id && msg.expect_response() {
-                    trace!(target: "network", ?msg, "initiate route back");
-                    self.routing_table_view.add_route_back(&clock, msg.hash(), my_peer_id);
-                }
-                self.tier2.send_message(peer_id, Arc::new(PeerMessage::Routed(msg)))
-            }
-            Err(find_route_error) => {
-                // TODO(MarX, #1369): Message is dropped here. Define policy for this case.
-                metrics::MessageDropped::NoRouteFound.inc(&msg.msg.body);
-
-                debug!(target: "network",
-                      account_id = ?self.config.validator.as_ref().map(|v|v.account_id()),
-                      to = ?msg.msg.target,
-                      reason = ?find_route_error,
-                      known_peers = ?self.routing_table_view.reachable_peers(),
-                      msg = ?msg.msg.body,
-                    "Drop signed message"
-                );
-                false
-            }
-        }
     }
 }
 
@@ -1258,7 +1036,10 @@ impl PeerManagerActor {
             peer_or_hash @ AccountOrPeerIdOrHash::PeerId(_)
             | peer_or_hash @ AccountOrPeerIdOrHash::Hash(_) => self.state.send_message_to_peer(
                 &self.clock,
-                RawRoutedMessage { target: peer_or_hash.clone(), body: msg },
+                self.state.sign_message(
+                    &self.clock,
+                    RawRoutedMessage { target: peer_or_hash.clone(), body: msg },
+                ),
             ),
         }
     }
@@ -1283,6 +1064,7 @@ impl PeerManagerActor {
         };
 
         let msg = RawRoutedMessage { target: AccountOrPeerIdOrHash::PeerId(target), body: msg };
+        let msg = self.state.sign_message(&self.clock, msg);
         if msg.body.is_important() {
             let mut success = false;
             for _ in 0..IMPORTANT_MESSAGE_RESENT_COUNT {
@@ -1433,7 +1215,10 @@ impl PeerManagerActor {
                 };
                 if self.state.send_message_to_peer(
                     &self.clock,
-                    RawRoutedMessage { target: AccountOrPeerIdOrHash::Hash(route_back), body },
+                    self.state.sign_message(
+                        &self.clock,
+                        RawRoutedMessage { target: AccountOrPeerIdOrHash::Hash(route_back), body },
+                    ),
                 ) {
                     NetworkResponses::NoResponse
                 } else {
@@ -1504,12 +1289,17 @@ impl PeerManagerActor {
                         {
                             if self.state.send_message_to_peer(
                                 &self.clock,
-                                RawRoutedMessage {
-                                    target: AccountOrPeerIdOrHash::PeerId(matching_peer.clone()),
-                                    body: RoutedMessageBody::PartialEncodedChunkRequest(
-                                        request.clone(),
-                                    ),
-                                },
+                                self.state.sign_message(
+                                    &self.clock,
+                                    RawRoutedMessage {
+                                        target: AccountOrPeerIdOrHash::PeerId(
+                                            matching_peer.clone(),
+                                        ),
+                                        body: RoutedMessageBody::PartialEncodedChunkRequest(
+                                            request.clone(),
+                                        ),
+                                    },
+                                ),
                             ) {
                                 success = true;
                                 break;
@@ -1530,10 +1320,13 @@ impl PeerManagerActor {
             NetworkRequests::PartialEncodedChunkResponse { route_back, response } => {
                 if self.state.send_message_to_peer(
                     &self.clock,
-                    RawRoutedMessage {
-                        target: AccountOrPeerIdOrHash::Hash(route_back),
-                        body: RoutedMessageBody::PartialEncodedChunkResponse(response),
-                    },
+                    self.state.sign_message(
+                        &self.clock,
+                        RawRoutedMessage {
+                            target: AccountOrPeerIdOrHash::Hash(route_back),
+                            body: RoutedMessageBody::PartialEncodedChunkResponse(response),
+                        },
+                    ),
                 ) {
                     NetworkResponses::NoResponse
                 } else {
@@ -1766,7 +1559,13 @@ impl PeerManagerActor {
                 trace!(target: "network", ?target, "Sending message to route back");
                 self.state.send_message_to_peer(
                     &self.clock,
-                    RawRoutedMessage { target: AccountOrPeerIdOrHash::Hash(target), body: *body },
+                    self.state.sign_message(
+                        &self.clock,
+                        RawRoutedMessage {
+                            target: AccountOrPeerIdOrHash::Hash(target),
+                            body: *body,
+                        },
+                    ),
                 );
                 PeerToManagerMsgResp::Empty
             }
