+++ chain/network/src/peer_manager/tests.rs
@@ -4,7 +4,8 @@ use crate::network_protocol::testonly as data;
 use crate::network_protocol::{Encoding, PeerAddr, SyncAccountsData};
 use crate::peer;
 use crate::peer_manager;
-use crate::peer_manager::peer_manager_actor::{Event as PME, LIMIT_PENDING_PEERS};
+use crate::peer_manager::network_state::LIMIT_PENDING_PEERS;
+use crate::peer_manager::peer_manager_actor::Event as PME;
 use crate::peer_manager::testonly::{Event, NormalAccountData};
 use crate::testonly::{assert_is_superset, make_rng, AsSet as _};
 use crate::types::{PeerMessage, RoutingTableUpdate};
